// src/api/identity/model/userRole.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * IdentityService API
 * Сервис аутентификации и авторизации платформы Scoodle.
 * OpenAPI spec version: v1
 */

/**
 * Поддерживаемые роли пользователя.
 */
export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  Admin: 'Admin',
  Teacher: 'Teacher',
  Student: 'Student',
} as const;